WDACConfig update 0.2.2

+++ WDACConfig/Deploy-SignedWDACConfig.psm1
@@ -24,6 +24,8 @@ function Deploy-SignedWDACConfig {
 
         [parameter(Mandatory = $false, ValueFromPipelineByPropertyName = $true)]
         [System.String]$SignToolPath,
+
+        [Parameter(Mandatory = $false)][Switch]$Deploy,
         
         [Parameter(Mandatory = $false)][Switch]$SkipVersionCheck
     )
@@ -132,26 +134,36 @@ function Deploy-SignedWDACConfig {
 
             Remove-Item ".\$PolicyID.cip" -Force            
             Rename-Item "$PolicyID.cip.p7" -NewName "$PolicyID.cip" -Force
-            CiTool --update-policy ".\$PolicyID.cip" -json | Out-Null
-            Write-Host "`npolicy with the following details has been Signed and Deployed in Enforced Mode:" -ForegroundColor Green        
-            Write-Output "PolicyName = $PolicyName"
-            Write-Output "PolicyGUID = $PolicyID`n"
-            Remove-Item -Path ".\$PolicyID.cip" -Force
 
-            # Ask user question about whether or not to add the Signed policy xml file to the User Config Json for easier usage later
-            $userInput = ''
-            while ($userInput -notin 1, 2) {
-                $userInput = $(Write-Host 'Add the Signed policy xml file path just created to the User Configurations? Please enter 1 to Confirm or 2 to Skip.' -ForegroundColor Cyan ; Read-Host) 
-                if ($userInput -eq 1) {
-                    Set-CommonWDACConfig -SignedPolicyPath $PolicyPath
-                    &$WriteViolet "Added $PolicyPath to the User Configuration file."             
+            if ($Deploy) {
+
+                CiTool --update-policy ".\$PolicyID.cip" -json | Out-Null
+                Write-Host "`npolicy with the following details has been Signed and Deployed in Enforced Mode:" -ForegroundColor Green        
+                Write-Output "PolicyName = $PolicyName"
+                Write-Output "PolicyGUID = $PolicyID`n"
+                Remove-Item -Path ".\$PolicyID.cip" -Force
+
+                # Ask user question about whether or not to add the Signed policy xml file to the User Config Json for easier usage later
+                $userInput = ''
+                while ($userInput -notin 1, 2) {
+                    $userInput = $(Write-Host 'Add the Signed policy xml file path just created to the User Configurations? Please enter 1 to Confirm or 2 to Skip.' -ForegroundColor Cyan ; Read-Host) 
+                    if ($userInput -eq 1) {
+                        Set-CommonWDACConfig -SignedPolicyPath $PolicyPath
+                        &$WriteHotPink "Added $PolicyPath to the User Configuration file."             
+                    }
+                    elseif ($userInput -eq 2) {                    
+                        &$WritePink 'Skipping...'                  
+                    }
+                    else {
+                        Write-Warning 'Invalid input. Please enter 1 or 2 only.'
+                    }               
                 }
-                elseif ($userInput -eq 2) {                    
-                    &$WritePink 'Skipping...'                  
-                }
-                else {
-                    Write-Warning 'Invalid input. Please enter 1 or 2 only.'
-                }               
+            }
+
+            else {            
+                Write-Host "`npolicy with the following details has been Signed and is ready for deployment:" -ForegroundColor Green
+                Write-Output "PolicyName = $PolicyName"
+                Write-Output "PolicyGUID = $PolicyID`n"
             }
         }
     }
@@ -184,6 +196,9 @@ Certificate common name
 .PARAMETER SignToolPath
 Path to the SignTool.exe - optional parameter
 
+.PARAMETER Deploy
+Indicates that the cmdlet will deploy the signed policy on the current system
+
 .PARAMETER SkipVersionCheck
 Can be used with any parameter to bypass the online version check - only to be used in rare cases
 
@@ -196,5 +211,5 @@ Can be used with any parameter to bypass the online version check - only to be u
 Set-PSReadLineKeyHandler -Key Tab -Function MenuComplete
 Register-ArgumentCompleter -CommandName 'Deploy-SignedWDACConfig' -ParameterName 'CertCN' -ScriptBlock $ArgumentCompleterCertificateCN
 Register-ArgumentCompleter -CommandName 'Deploy-SignedWDACConfig' -ParameterName 'PolicyPaths' -ScriptBlock $ArgumentCompleterPolicyPaths
-Register-ArgumentCompleter -CommandName 'Deploy-SignedWDACConfig' -ParameterName 'CertPath' -ScriptBlock $ArgumentCompleterCertPath
-Register-ArgumentCompleter -CommandName 'Deploy-SignedWDACConfig' -ParameterName 'SignToolPath' -ScriptBlock $ArgumentCompleterSignToolPath
+Register-ArgumentCompleter -CommandName 'Deploy-SignedWDACConfig' -ParameterName 'CertPath' -ScriptBlock $ArgumentCompleterCerFilePathsPicker
+Register-ArgumentCompleter -CommandName 'Deploy-SignedWDACConfig' -ParameterName 'SignToolPath' -ScriptBlock $ArgumentCompleterExeFilePathsPicker
